chore: added unauthenticated navbar
