+++ sb-web/src/modules/navbar/Navbar.scss
@@ -7,8 +7,8 @@
     z-index: 99;
 
     .logo {
-        height: 30px !important;
-        width: 30px !important;
+        height: 35px !important;
+        width: 35px !important;
         z-index: 98;
 
         &:hover {
