+++ sb-web/src/screens/home/Home.tsx
@@ -1,5 +1,7 @@
 import React, { FunctionComponent } from 'react';
 
+import { Box } from '@chakra-ui/layout';
+
 export const Home: FunctionComponent = () => {
-    return <div>Home Page</div>;
+    return <Box px={{ base: '4', sm: '8' }}>Home Page</Box>;
 };
