+++ sb-web/src/util/Icons.ts
@@ -22,4 +22,9 @@ export { CgMenuRight as Menu } from 'react-icons/cg';
 export { FcGoogle as Google } from 'react-icons/fc';
 export { FiPlus as Plus, FiMinus as Minus, FiCheck as Checkmark } from 'react-icons/fi';
 export { BiErrorCircle as ErrorCircle } from 'react-icons/bi';
+export { VscBell as Bell } from 'react-icons/vsc';
+export { MdOutlineImageSearch as Memories } from 'react-icons/md';
+
 export { ReactComponent as Boat } from 'assets/icons/boat.svg';
+export { ReactComponent as Logo } from 'assets/sailboat-logo.svg';
+export { ReactComponent as LogoType } from 'assets/icon-logo-type.svg';
